0-linear.c

+++ 0x1E-search_algorithms/0-linear.c
@@ -9,20 +9,18 @@
 
 int linear_search(int *array, size_t size, int value)
 {
-	unsigned int i = 0;
+	size_t i = 0;
 
 	if (array == NULL)
 		return (-1);
 
 	while (i < size)
-		{
-			printf("Value checked array[%d] = [%d]\n", i, array[i]);
-
-			if (value == array[i])
-				return (i);
-
-			i++;
-		}
+	{
+		printf("Value checked array[%ld] = [%d]\n", i, array[i]);
+		if (value == array[i])
+			return (i);
+		i++;
+	}
 
 	return (-1);
 }
